thinkpolls/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.1.1
django-cors-headers==3.0.2
djangorestframework==3.10.1
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.0
idna==2.8
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
